Cleaned up the choosing of the default CONTAINER_IFNAME which is now set further down to allow easier defaulting for infiniband.
Set default IPoIB interface to ib0.  I realised that the reason eth1 was
chosen for ethernet was because eth0 is the default for default
docker networking.  It actually makes more sense to set the infiniband iboip
interface to the first infiniband device or ib0 and not ib1

Don't allow a mac address to be set on an IPoIB interface.  It can't be done.

Moved WAIT check further down to a position where CONTAINER_IFNAME is now
correctly set.  So now it allows pipework to wait for an infiniband interface.

Updated README.md to contain comments about infiniband IPoIB.