Fix loose grep match
When determining whether a port needs to be added to an ovs switch, pipework
first greps to see if a port by that name is on the switch, in which case it
skips adding the port.

The check uses a grep which also matches against ports which contain the name
to be added as a substring, erroneously determining that the port is already on
the switch.

This change tightens the grep to look for exactly the desired port name.