Adding metadata to openvswitch database.
Pipework doesn't populate the openvswitch database with any information
about the pipework interface creation. This makes it difficult to glean
any information about the container to which an interface is assigned.

This patch adds the bridge, interface name, container ID and vlan to
the database so other tools can query it.

This patch was specifically designed to help tools like dreamcat4's
docker image clean up these interfaces when the container is started
with something like 'docker run --rm'.